refactor: 불필요한 팩토리 클래스 제거, 테스트 코드에서만 사용하는 코드 프로덕션 코드에서 제거
