+++ src/test/java/nextstep/ladder/ladder/LadderTest.java
@@ -1,6 +1,6 @@
 package nextstep.ladder.ladder;
 
-import nextstep.ladder.factory.LadderFactory;
+import nextstep.ladder.ladder.stub.AlwaysAddLadderStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +11,7 @@ class LadderTest {
     @DisplayName("사다리 높이 만큼 사다리를 생성한다.")
     @Test
     void create_ladder() {
-        Ladder ladder = Ladder.of(LadderBound.of(4, 5), LadderFactory.alwaysAddLadderStrategy());
+        Ladder ladder = Ladder.of(LadderBound.of(4, 5), new AlwaysAddLadderStrategy());
 
         assertThat(ladder.getLines()).hasSize(5);
     }
