+++ src/test/java/nextstep/ladder/ladder/stub/AlwaysAddLadderStrategy.java
@@ -1,4 +1,6 @@
-package nextstep.ladder.strategy;
+package nextstep.ladder.ladder.stub;
+
+import nextstep.ladder.strategy.LadderStrategy;
 
 public class AlwaysAddLadderStrategy implements LadderStrategy {
     @Override
